+++ workers/unity/Assets/Fps/Scripts/GameLogic/Respawning/SpawnPoints/SpawnPointIndicator.cs
@@ -4,11 +4,12 @@ namespace Fps
 {
     public class SpawnPointIndicator : MonoBehaviour
     {
-        private readonly Vector3 boxDimensions = new Vector3(1, 0, 1);
+        private static readonly Color gizmoColor = new Color(.1f, 1f, .1f, .8f); // Transparent green
+        private static readonly Vector3 boxDimensions = new Vector3(.5f, 2f, .5f);
 
         private void OnDrawGizmos()
         {
-            Gizmos.color = Color.blue;
+            Gizmos.color = gizmoColor;
             var position = transform.position;
             var rotation = Quaternion.Euler(0, transform.eulerAngles.y, 0);
             Gizmos.matrix = Matrix4x4.TRS(
@@ -17,8 +18,8 @@ namespace Fps
                 Vector3.one
             );
 
-            Gizmos.DrawWireCube(Vector3.zero, boxDimensions);
-            Gizmos.DrawLine(Vector3.zero, Vector3.forward * boxDimensions.z / 2);
+            Gizmos.DrawCube(Vector3.up * boxDimensions.y * .5f, boxDimensions);
+            Gizmos.DrawLine(Vector3.forward * boxDimensions.z * .5f, Vector3.forward * boxDimensions.z);
         }
     }
 }