+++ CHANGELOG.md
@@ -8,11 +8,21 @@
     - Added ability to choose deployment region (US, EU).
     - Added ability to force-upload an assembly.
     - Automatically open the SpatialOS Console page for a launched deployment.
+- New way to control tile distribution and types in the world via TileTypeVolume component:
+    - Create a collection of tile types with a TileTypeCollection asset (Assets > Create > Improbable > Tile Type Collection).
+    - Place GOs with a TileTypeVolume component into a prefab to control where to spawn those tiles in the world.
+    - Specify the above TileTypeVolumes prefab to use inside MapBuilderSettings config object.
+    - Specify a 'default' TileTypeCollection to use in the MapBuilderSettings to use if no volume found at a tile location.
+    - Preexisting tile types: 'Default', 'Mountain', 'Residential', 'Structure', 'Tower', 'Wild'
 
 ### Changed
 
 - Reduced `fps_simulated_players_creation_interval` from 60 to 5.
 - Disabled the `Generate Map` button in the MapBuilder window if `MapBuilderSettings` is not set.
+- The four middle tiles of the world are no longer forced to be of certain tile type.
+- World tile prefabs improved.
+	- Now make use of nested prefabs.
+	- Completely new set of default tiles used to populate the world
 - Updated the build configuration asset
 
 ### Fixed